+++ lib/hycop/database/firebase_database.dart
@@ -81,6 +81,29 @@ class FirebaseDatabase extends AbsDatabase {
   }
 
   @override
+  Future<void> updateData(String collectionId, String mid, Map<dynamic, dynamic> data) async {
+    await initialize();
+
+    // Map<String, Object?> converted = {};
+    // for (MapEntry e in data.entries) {
+    //   if (e.value is List) {
+    //     converted[e.key.toString()] = FieldValue.arrayUnion(e.value);
+    //   } else {
+    //     converted[e.key.toString()] = e.value as Object?;
+    //   }
+    // }
+    CollectionReference collectionRef = _db!.collection(collectionId);
+    //await collectionRef.doc(mid).update(converted);
+    await collectionRef.doc(mid).set(data, SetOptions(merge: true));
+    // for (MapEntry e in data.entries) {
+    //   if (e.value is List) {
+    //     await collectionRef.doc(mid).update({e.key.toString(): FieldValue.arrayUnion(e.value)});
+    //   }
+    // }
+    logger.finest('$mid saved');
+  }
+
+  @override
   Future<void> createData(String collectionId, String mid, Map<dynamic, dynamic> data) async {
     await initialize();